subnet/validator/requirements.txt
bittensor
torch
flask
pydantic
python-dotenv
simplejson
starlette
uvicorn
requests-async
